Enable actor data race detection and fix resulting problems
This revealed a number of places where callbacks were invoked on a background
thread but weren't marked as Sendable. It also unforunately revealed some more
compiler bugs related to property wrappers and sendability; the
`projectedValue` hack with `@Locked` doesn't work with data race detection
enabled because it generates an actor-confined closure as part of the
implementation.